adding samples generated -refs #9203

+++ samples/server/petstore/jaxrs-cxf-non-spring-app/src/main/java/io/swagger/api/impl/PetApiServiceImpl.java
@@ -38,7 +38,7 @@ public class PetApiServiceImpl implements PetApi {
      * Deletes a pet
      *
      */
-    public void deletePet(Integer petId, String apiKey) {
+    public void deletePet(Long petId, String apiKey) {
         // TODO: Implement...
         
         
@@ -74,7 +74,7 @@ public class PetApiServiceImpl implements PetApi {
      * Returns a single pet
      *
      */
-    public Pet getPetById(Integer petId) {
+    public Pet getPetById(Long petId) {
         // TODO: Implement...
         
         return null;
@@ -94,7 +94,7 @@ public class PetApiServiceImpl implements PetApi {
      * Updates a pet in the store with form data
      *
      */
-    public void updatePetWithForm(Integer petId, String name, String status) {
+    public void updatePetWithForm(Long petId, String name, String status) {
         // TODO: Implement...
         
         
@@ -104,7 +104,7 @@ public class PetApiServiceImpl implements PetApi {
      * uploads an image
      *
      */
-    public ModelApiResponse uploadFile(Integer petId, String additionalMetadata,  Attachment fileDetail) {
+    public ModelApiResponse uploadFile(Long petId, String additionalMetadata,  Attachment fileDetail) {
         // TODO: Implement...
         
         return null;
